Remove old version of BatchQueue
